refactor: change TokenContext to withToken, update readme
